Cursed Possessions Update!
It has been confirmed that the ghost aggression levels heard on the radio are only for flavour text, so, why not fill the space with something more useful? I've swapped out the ghost aggression text for the new Cursed Possessions! Information taken from the wiki on each of the items.